+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Mapbox welcomes participation and contributions from everyone.
 
 ## main
+* [tile store] Expose API for estimating Tile Region downloads and storage size.
 
 # 11.3.0-beta.1 - 14 March, 2024