+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+[tile store] Expose API for estimating Tile Region downloads and storage size.
+
 # 11.3.0-beta.1 - 14 March, 2024
 
 * Update the minimum Xcode version to 15.2 (Swift 5.9).